refactor(architecture): simplify and focus project for imitation learning
- Renamed all `*Config` classes to `*Model` and organized them into domain-specific files under `src/configs/models/` for clarity.
- Replaced the orchestrator with a clean `train_imitation_learning()` function and a focused `save_checkpoint()` utility, removing unnecessary layers of abstraction.
- Introduced a dedicated entry point for imitation learning (`src/configs/entry_points/imitation_train.py`), ensuring that only necessary components are configured.
- The new `src/thermur/scripts/train_imitation.py` script now directly instantiates components and calls the training function, simplifying the execution flow.
- All Hydra-zen builds have been updated to use the new model locations and naming conventions, and unnecessary builds were removed to reduce clutter.
- Deleted numerous old files, including the orchestrator, old training configurations, and unused build files, to align with the new, simplified architecture.
