+++ REFACTORING_SUMMARY.md
@@ -1,0 +1,95 @@
+# Thermur Project Refactoring Summary
+
+## Completed Refactoring
+
+### 1. Pydantic Models Reorganization ✅
+- Renamed all `*Config` classes to `*Model` to accurately reflect their purpose
+- Organized models into domain-specific files in `src/configs/models/`:
+  - `agent.py`: AgentModel, SwarmModel
+  - `environment.py`: EnvironmentModel, ExpertPolicyModel
+  - `policy.py`: GNNModel
+  - `safety.py`: CBFModel, QPSolverModel
+  - `training.py`: HyperparameterModel, CollectorModel, ReplayBufferModel, CheckpointModel
+  - `logging.py`: LoggingModel, WandbModel
+
+### 2. Entry Point Configuration ✅
+- Created `src/configs/entry_points/imitation_train.py` with focused config for imitation learning
+- Only includes components necessary for training (no orchestrator, no irrelevant components)
+- Added proper `__init__.py` for entry points package
+
+### 3. Training Utilities ✅
+- Created `src/thermur/training/imitation.py` with clean `train_imitation_learning()` function
+- Moved `ImitationLoss` to `src/thermur/training/loss.py`
+- Removed orchestrator pattern - training logic is now a simple function
+- Added focused `save_checkpoint()` utility
+
+### 4. Clean Entry Point Script ✅
+- Created `src/thermur/scripts/train_imitation.py` 
+- Directly instantiates components using Hydra
+- Calls training function without orchestrator overhead
+
+### 5. Updated All Builds ✅
+- Updated all build files to import from new model locations (`configs.models.*`)
+- Updated to use new model names (e.g., `EnvironmentModel` instead of `EnvironmentConfig`)
+- Ensured consistent naming for `cls_name` in `zen_dataclass` (using `*Build` suffix)
+- Removed unnecessary builds not used for imitation learning
+
+### 6. Cleaned Up Old Files ✅
+- Removed `src/configs/pydantic.py` (replaced by domain-specific model files)
+- Removed `src/configs/train_config.py` (replaced by entry point configs)
+- Removed `src/thermur/core/orchestrator.py` (replaced by training utilities)
+- Removed `src/thermur/scripts/train.py` (replaced by train_imitation.py)
+- Removed `src/configs/builds/orchestrator.py` (no longer needed)
+- Removed `src/configs/builds/train.py` (consolidated into entry point)
+- Removed `src/configs/builds/safety.py` (not used in imitation learning)
+
+### 7. Updated All Imports ✅
+- Updated `src/configs/__init__.py` to export only entry point registration
+- Updated `src/thermur/__init__.py` to remove SafetyFilter and add ImitationLoss
+- Updated `src/thermur/core/__init__.py` to remove orchestrator imports
+- Updated all `__init__.py` files to reflect new structure
+
+## Current Architecture
+
+The project now follows a clean, focused architecture for imitation learning:
+
+```
+Pydantic Models (src/configs/models/)
+    ↓
+Hydra-zen Builds (src/configs/builds/)
+    ↓
+Entry Point Config (src/configs/entry_points/)
+    ↓
+Training Script (src/thermur/scripts/train_imitation.py)
+    ↓
+Training Function (src/thermur/training/imitation.py)
+```
+
+## Benefits of New Architecture
+
+1. **Clear separation of concerns**: Models, builds, and entry points are clearly separated
+2. **Focused configurations**: Only includes components needed for imitation learning
+3. **Direct instantiation**: No orchestrator indirection, cleaner code flow
+4. **Type safety**: Pydantic models ensure type checking throughout
+5. **Simplicity**: Removed all redundant code and approaches
+
+## Future Extensions
+
+When adding new functionality:
+- Create new entry point configs in `src/configs/entry_points/`
+- Add corresponding scripts in `src/thermur/scripts/`
+- Reuse existing models and builds where possible
+- Keep each entry point focused on its specific purpose
+
+Got it. Here is the commit message with plain bullet points.
+
+```
+refactor(architecture): simplify and focus project for imitation learning
+
+- Renamed all `*Config` classes to `*Model` and organized them into domain-specific files under `src/configs/models/` for clarity.
+- Replaced the orchestrator with a clean `train_imitation_learning()` function and a focused `save_checkpoint()` utility, removing unnecessary layers of abstraction.
+- Introduced a dedicated entry point for imitation learning (`src/configs/entry_points/imitation_train.py`), ensuring that only necessary components are configured.
+- The new `src/thermur/scripts/train_imitation.py` script now directly instantiates components and calls the training function, simplifying the execution flow.
+- All Hydra-zen builds have been updated to use the new model locations and naming conventions, and unnecessary builds were removed to reduce clutter.
+- Deleted numerous old files, including the orchestrator, old training configurations, and unused build files, to align with the new, simplified architecture.
+```
