+++ src/configs/__init__.py
@@ -1,29 +1,9 @@
 """
 Configuration management for the Thermur project.
 
-This package provides Pydantic models and Hydra-zen builders for declarative
-configuration of all system components.
+This package provides Pydantic models and Hydra-zen builders for
+imitation learning training configuration.
 """
-from configs.app      import register_configs
-from configs.pydantic import *
+from .entry_points import register_imitation_training_config
 
-__all__ = [
-    "register_configs",
-    
-    # From pydantic (using *)
-    "AgentConfig",
-    "CBFConfig",
-    "CheckpointConfig",
-    "CollectorConfig",
-    "EnvironmentConfig",
-    "ExpertPolicyConfig",
-    "GNNConfig",
-    "LoggingConfig",
-    "PolicyConfig",
-    "QPSolverConfig",
-    "ReplayBufferConfig",
-    "SafetyConfig",
-    "SwarmConfig",
-    "TrainConfig",
-    "WandbConfig",
-]
+__all__ = ["register_imitation_training_config"]
