+++ src/configs/builds/loss.py
@@ -4,17 +4,17 @@ Hydra-zen builder for the imitation learning loss module.
 This module defines the configuration builder for ImitationLoss, which
 computes the MSE between predicted and expert actions for behavioral cloning.
 """
-from hydra_zen import builds
-from omegaconf import SI
-from thermur   import ImitationLoss
+from hydra_zen         import builds
+from omegaconf         import SI
+from thermur.training  import ImitationLoss
 
 
 build_loss = builds(
     ImitationLoss,
-    policy_network          = SI("${gnn_policy}"),  # Reference to instantiated GNN policy
+    policy_network          = SI("${policy}"),
     populate_full_signature = True,
     zen_dataclass           = {
         "module"   : "src.configs.builds.loss",
-        "cls_name" : "LossConfig"
+        "cls_name" : "LossBuild"
     }
 )
