+++ src/configs/builds/policy.py
@@ -5,35 +5,31 @@ This module defines configuration builders for both the expert flocking
 controller and the GNN policy that learns from it. These builders leverage
 Pydantic validation through the zen() wrapper.
 """
-from configs    import AgentConfig, ExpertPolicyConfig, GNNConfig
-from hydra_zen  import builds, zen
-from thermur    import ExpertFlockingController, GNNPolicy
+from configs.models import AgentModel, ExpertPolicyModel, GNNModel
+from hydra_zen      import builds, zen
+from thermur        import ExpertFlockingController, GNNPolicy
 
 
 # Expert controller that generates demonstration data
 build_expert_controller = builds(
     ExpertFlockingController,
-    expert_config           = zen(ExpertPolicyConfig),
-    agent_config            = zen(AgentConfig),
+    expert_config           = zen(ExpertPolicyModel),
+    agent_config            = zen(AgentModel),
     populate_full_signature = True,
     zen_dataclass           = {
         "module"   : "src.configs.builds.policy",
-        "cls_name" : "ExpertControllerConfig"
+        "cls_name" : "ExpertControllerBuild"
     }
 )
 
 
 # GNN policy that learns from expert demonstrations
-build_gnn_policy = builds(
+build_policy = builds(
     GNNPolicy,
-    config                  = zen(GNNConfig),
+    config                  = zen(GNNModel),
     populate_full_signature = True,
     zen_dataclass           = {
         "module"   : "src.configs.builds.policy", 
-        "cls_name" : "GNNPolicyConfig"
+        "cls_name" : "PolicyBuild"
     }
 )
-
-
-# For convenience, also export as expert policy
-build_expert_policy = build_expert_controller
