+++ src/thermur/models/__init__.py
@@ -3,6 +3,6 @@ Neural network models for the Thermur project.
 
 This package contains the Graph Neural Network policy and related models.
 """
-from thermur.models.gnn_policy import GNNPolicy
+from .gnn_policy import GNNPolicy
 
 __all__ = ["GNNPolicy"]
