+++ src/thermur/ops/__init__.py
@@ -3,8 +3,8 @@ Operational utilities for the Thermur project.
 
 This package contains data sources, logging configuration, and seed management.
 """
-from thermur.ops.data   import EnvironmentDataSource
-from thermur.ops.loguru import configure_loguru
-from thermur.ops.seed   import set_seed
+from .data   import EnvironmentDataSource
+from .loguru import configure_loguru
+from .seed   import set_seed
 
 __all__ = ["configure_loguru", "EnvironmentDataSource", "set_seed"]
