+++ src/thermur/physics/__init__.py
@@ -4,7 +4,7 @@ Physics-based controllers and safety filters for the Thermur project.
 This package contains the expert flocking controller and Control Barrier
 Function safety filter.
 """
-from thermur.physics.potentials import ExpertFlockingController
-from thermur.physics.safety     import SafetyFilter
+from .potentials import ExpertFlockingController
+from .safety     import SafetyFilter
 
 __all__ = ["ExpertFlockingController", "SafetyFilter"]
