+++ src/thermur/training/__init__.py
@@ -1,0 +1,15 @@
+"""
+Training utilities for the Thermur project.
+
+This package provides clean, focused training functions
+without the overhead of orchestrator classes.
+"""
+from .imitation import save_checkpoint, train_imitation_learning
+from .loss      import ImitationLoss
+
+
+__all__ = [
+    "train_imitation_learning",
+    "save_checkpoint",
+    "ImitationLoss",
+]
